requirements.txt
# Audio processing
pydub>=0.25.1
noisereduce>=3.0.0
librosa>=0.10.0
scipy>=1.10.0
numpy>=1.24,<2.0

# Speaker diarization
pyannote.audio>=3.1.0,<4.0
soundfile>=0.12.0

# Transcription — faster-whisper (CTranslate2-based, GPU-accelerated)
faster-whisper>=1.0.0

# HuggingFace hub — must stay <1.0.0 for pyannote 3.x internal API compatibility
huggingface_hub>=0.16.0,<1.0.0

# PyTorch — CUDA 12.1 build for NVIDIA GPU (GTX 1650 / sm_75)
# Install with: pip install torch==2.1.0+cu121 torchaudio==2.1.0+cu121 --index-url https://download.pytorch.org/whl/cu121
# For CPU-only fallback: pip install torch==2.1.0+cpu torchaudio==2.1.0+cpu --index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cu121
torchaudio==2.1.0+cu121

# Config / utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Claude API — Stage 5 report generation (optional; only used with --report flag)
anthropic>=0.40.0
